executive summary restructure
